requirement.txt
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-telegram-bot==21.4
sniffio==1.3.1
Werkzeug==3.0.4
